Used to automatically Log you into sonicwall.
This is the the first script I wrote when I started working on a way to
somehow automate the whole process of logging into the sonicwall, it
seems tiresome how I had to keep logging myself in, and also this
hampered downloads which took long periods of time.
In mobiles especially it was a pain to Log in.
I could implement the run the script again after 2 and a half hours part
which was usefull, but since in my college the logging in to sonicwall
is no longer required in order to access the Internet.
I will be working on this later when/If they bring back the sonicwall.

Signed-off-by: Aditya Prasad <adityaprasadiskool@gmail.com>