Add proof of working for Kruskal's.py and Update Readme
I added the .txt file showing the I/O of the test with a
graph of 21 edges and 10 vertex's. And this graph and its
MST is shown in the .jpg, both named Kruskal_test.
I removed the query for the maximum value of the edge weights
As we are not using counting sort. If needed we should find it
from the edges that we recive.
I updated the README.md with information about the last two scripts
that I added to the repo.

Signed-off-by: Aditya Prasad <adityaprasadiskool@gmail.com>