Script to scrape and save NITT student information
The previous commit was the file when I used javascript to emulate
the user, now I am sticking with selenium, which does it job very well!
I used javascript when I worked with the face of the results page,
now that I found the actual url, I have direct access to the
elements in the page.
I have written extensive comments for the main statements.

Signed-off-by: Aditya Prasad <adityaprasadiskool@gmail.com>