Fix NoSuchElementException when unused roll number is encountered
There are some roll numbers which are assigned to students and these
students leave the college etc, Anyway, when such roll numbers which
have no ACTIVE student is entered into the roll number field, the drop
down menu from which we can choose the second sem option and get the
results to appear on the webpage.
So I caught this error, and handled it.
I added more comments to explain it all.
I removed the unnecessary import which should have been done
in the last commit, my bad.
I made the message telling the user the status of the scraping
more informative.

Signed-off-by: Aditya Prasad <adityaprasadiskool@gmail.com>