Update Readme and FIX scraper status message
I updated the readme with this latest script(scrape_results)'s
information.
I also fixed the status message which is displayed after each
student's data is saved to the disk. The slicing I used on
the main string to generate the message didnt work when I ran the
script targeting a different department. So I have now generilized
it and it should make sense for most if not all departments.
It doesnt really matter anyway.

Signed-off-by: Aditya Prasad <adityaprasadiskool@gmail.com>