Add bisection method and modify scrape_results for 3rd sem results
I have made this new script which implements the Bisection  method
sir taught today in class. It is used to approximate the root of
a polynomial (which for now has to be defined in the program itself)
using bisection method which as you can see is similar to binary search.

I also modified the scraper to fetch the 3rd semester marks as the text
on the drop down box has changed.

Signed-off-by: Aditya Prasad <adityaprasadiskool@gmail.com>