Implemented Regula-Falsi method and close a issue in a different repo
I have reused most of the code used in the bisection method, i feel
like i should use function for the code which are being repeated for
both methods, for know the method is not working to a reasonable level
of accuracy.

I am also going to try to close a issue in my delta repository via this
commit. Thus this closes Aditya8795/Delta#1

Signed-off-by: Aditya Prasad <adityaprasadiskool@gmail.com>