Adding 2 scripts in progress
I have almost given up on kissanime, its protected by cloudflare.
I may try later but the code can right now open the episodes in
various windows.

The gaussian elimination is in the middle of implementation, I am trying
to manipulate the python lists as a 2D array. I will now shift to using
nested lists for it will be much easier to manipulate that way.

Signed-off-by: Aditya Prasad <adityaprasadiskool@gmail.com>