Finished Script
I guess when I updated all my python pakages
my selenium got to cheat the cloudflare server.

The delay of 10 seconds is a little steep. Ill try to find
a different way, I tried the following

This didnt work as there about 5 seconds during which CloudFlare transfers
us to the actual site. so "driver.find_element_by_id('head')" wont work

from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.common.exceptions import TimeoutException
delay = 10 # seconds
try:
    WebDriverWait(driver, delay).until(EC.presence_of_element_located(driver.find_element_by_id('head')))
    print "Page is ready!"
except TimeoutException:
    print "Loading took too much time!"

the point is we need to wait for some elements of the page to load.
This should be possible, we dont even need to wait for all the ajax requests to
complete. the adverts and all take so much time.

A work around I found is to stop the loading in the firefox browser after it seems to have loaded.
I encountered cases where the page seemed to be loading for ever!!

Other than that, the blogspot links shouldnt be taken, the option of choosing anime
and quailty of download should be added.

Signed-off-by: Aditya A Prasad <adityaprasadiskool@gmail.com>